Money Manager 24.7.2
* Обновление зависимостей
* Загрузка брокерского отчёта Сбера из HTML
* Временно отключена загрузка брокерского отчёта Сбера из Excel
* Ручной ввод сплита ценной бумаги
* Перерасчёт стоимости портфеля с учётом сплитов

+++ desktop/src/main/java/org/panteleyev/money/app/investment/InvestmentSummary.java
@@ -10,13 +10,13 @@ import java.util.UUID;
 public record InvestmentSummary(
         UUID securityUuid,
         BigDecimal averagePrice,
-        int securityAmount,
+        BigDecimal securityAmount,
         BigDecimal totalValue,
         BigDecimal percentage,
         BigDecimal totalExchangeFee,
         BigDecimal totalBrokerFee,
         // Доп. данные для рассчета
-        int totalPurchaseAmount,
+        BigDecimal totalPurchaseAmount,
         BigDecimal totalPurchaseValue
 ) {
     public InvestmentSummary withPercentage(BigDecimal percentage) {
